refactor package of route module
