+++ shardingsphere-underlying/shardingsphere-route/src/main/java/org/apache/shardingsphere/underlying/route/decorator/DateNodeRouteDecorator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.underlying.route;
+package org.apache.shardingsphere.underlying.route.decorator;
 
 import org.apache.shardingsphere.underlying.route.context.RouteContext;
